Modified few models, controllers, and some view

+++ database/migrations/2024_08_02_035720_create_dispen_siswa_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
     {
         Schema::create('dispen_siswa', function (Blueprint $table) {
             $table->unsignedBigInteger('id_dispen');
-            $table->unsignedBigInteger('id_siswa');
+            $table->integer('id_siswa');
 
             $table->primary(['id_dispen', 'id_siswa']);
             $table->foreign('id_dispen')->references('id_dispen')->on('dispen');
